requirements.txt
mysql-connector-python==8.0.20
pkg-resources==0.0.0
prettytable==0.7.2
protobuf==3.11.3
six==1.14.0
